+++ src/main/python/plugin_text.py
@@ -15,7 +15,10 @@ limitations under the License.
 '''
 
 from fmt_template import Formatter
+from ofexport import load_template, format_document
 
-class PrintTextVisitor(Formatter):
-    def __init__ (self, out, template):
-        Formatter.__init__(self, out, template)
+def generate (out, root_project, root_context, project_mode, template_dir, type_config):
+    subject = root_project if project_mode else root_context
+    template = load_template (template_dir, type_config['template'])
+    visitor = Formatter (out, template)
+    format_document (subject, visitor, project_mode)